Move TerrainManifest
