+++ emergence_lib/src/asset_management/ui.rs
@@ -7,12 +7,10 @@ use core::hash::Hash;
 use crate::{
     player_interaction::terraform::TerraformingChoice,
     structures::structure_manifest::{Structure, StructureManifest},
+    terrain::terrain_manifest::TerrainManifest,
 };
 
-use super::{
-    manifest::{Id, TerrainManifest},
-    Loadable,
-};
+use super::{manifest::Id, Loadable};
 
 /// Stores all structural elements of the UI: buttons, frames, widgets and so on
 #[derive(Resource)]
