+++ emergence_lib/src/infovis.rs
@@ -10,14 +10,12 @@ use bevy::{
 use core::fmt::Display;
 
 use crate::{
-    asset_management::{
-        manifest::{Id, Terrain},
-        terrain::TerrainHandles,
-    },
+    asset_management::{manifest::Id, terrain::TerrainHandles},
     enum_iter::IterableEnum,
     player_interaction::{selection::ObjectInteraction, InteractionSystem},
     signals::{SignalKind, SignalStrength, SignalType, Signals},
     simulation::geometry::TilePos,
+    terrain::terrain_manifest::Terrain,
     units::unit_manifest::Unit,
 };
