+++ emergence_lib/src/player_interaction/cursor.rs
@@ -6,9 +6,10 @@ use leafwing_input_manager::prelude::ActionState;
 
 use super::{InteractionSystem, PlayerAction};
 use crate::{
-    asset_management::manifest::{Id, Terrain},
+    asset_management::manifest::Id,
     simulation::geometry::TilePos,
     structures::{construction::Ghost, structure_manifest::Structure},
+    terrain::terrain_manifest::Terrain,
     units::unit_manifest::Unit,
 };
