+++ emergence_lib/src/player_interaction/terraform.rs
@@ -3,12 +3,10 @@
 use bevy::prelude::*;
 
 use crate::{
-    asset_management::{
-        manifest::{Id, Terrain, TerrainManifest},
-        terrain::TerrainHandles,
-    },
+    asset_management::{manifest::Id, terrain::TerrainHandles},
     simulation::geometry::{Height, TilePos},
     structures::commands::StructureCommandsExt,
+    terrain::terrain_manifest::{Terrain, TerrainManifest},
 };
 
 use super::{zoning::Zoning, InteractionSystem};
