+++ emergence_lib/src/simulation/generation.rs
@@ -1,11 +1,12 @@
 //! Generating starting terrain and organisms
-use crate::asset_management::manifest::{Id, Terrain};
+use crate::asset_management::manifest::Id;
 use crate::asset_management::units::UnitHandles;
 use crate::asset_management::AssetState;
 use crate::player_interaction::clipboard::ClipboardData;
 use crate::simulation::geometry::{Facing, Height, TilePos};
 use crate::structures::commands::StructureCommandsExt;
 use crate::structures::structure_manifest::StructureManifest;
+use crate::terrain::terrain_manifest::Terrain;
 use crate::terrain::SpawnTerrainCommand;
 use crate::units::unit_manifest::UnitManifest;
 use crate::units::UnitBundle;
